move server and analyzer out of main
