+++ pkg/testgridanalysis/testgridhelpers/testgridhelpers.go
@@ -1,0 +1,154 @@
+package testgridhelpers
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"io"
+	"io/ioutil"
+	"net/http"
+	"os"
+	"regexp"
+	"strings"
+	"time"
+
+	testgridv1 "github.com/openshift/sippy/pkg/apis/testgrid/v1"
+	"github.com/openshift/sippy/pkg/util"
+	"k8s.io/klog"
+)
+
+var (
+	DashboardTemplate = "redhat-openshift-ocp-release-%s-%s"
+)
+
+func DownloadData(releases []string, filter string, storagePath string) {
+	var jobFilter *regexp.Regexp
+	if len(filter) > 0 {
+		jobFilter = regexp.MustCompile(filter)
+	}
+
+	for _, release := range releases {
+
+		dashboard := fmt.Sprintf(DashboardTemplate, release, "blocking")
+		err := downloadJobSummaries(dashboard, storagePath)
+		if err != nil {
+			klog.Errorf("Error fetching dashboard page %s: %v\n", dashboard, err)
+			continue
+		}
+		blockingJobs, _, err := LoadJobSummaries(dashboard, storagePath)
+		if err != nil {
+			klog.Errorf("Error loading dashboard page %s: %v\n", dashboard, err)
+			continue
+		}
+
+		for jobName, job := range blockingJobs {
+			if util.RelevantJob(jobName, job.OverallStatus, jobFilter) {
+				klog.V(4).Infof("Job %s has bad status %s\n", jobName, job.OverallStatus)
+				err := downloadJobDetails(dashboard, jobName, storagePath)
+				if err != nil {
+					klog.Errorf("Error fetching job details for %s: %v\n", jobName, err)
+				}
+			}
+		}
+	}
+
+	for _, release := range releases {
+
+		dashboard := fmt.Sprintf(DashboardTemplate, release, "informing")
+		err := downloadJobSummaries(dashboard, storagePath)
+		if err != nil {
+			klog.Errorf("Error fetching dashboard page %s: %v\n", dashboard, err)
+			continue
+		}
+		informingJobs, _, err := LoadJobSummaries(dashboard, storagePath)
+		if err != nil {
+			klog.Errorf("Error fetching dashboard page %s: %v\n", dashboard, err)
+			continue
+		}
+
+		for jobName, job := range informingJobs {
+			if util.RelevantJob(jobName, job.OverallStatus, jobFilter) {
+				klog.V(4).Infof("Job %s has bad status %s\n", jobName, job.OverallStatus)
+				err := downloadJobDetails(dashboard, jobName, storagePath)
+				if err != nil {
+					klog.Errorf("Error fetching job details for %s: %v\n", jobName, err)
+				}
+			}
+		}
+	}
+}
+
+func LoadJobSummaries(dashboard string, storagePath string) (map[string]testgridv1.JobSummary, time.Time, error) {
+	jobs := make(map[string]testgridv1.JobSummary)
+	url := fmt.Sprintf("https://testgrid.k8s.io/%s/summary", dashboard)
+
+	var buf *bytes.Buffer
+	filename := storagePath + "/" + "\"" + strings.ReplaceAll(url, "/", "-") + "\""
+	b, err := ioutil.ReadFile(filename)
+	if err != nil {
+		return jobs, time.Time{}, fmt.Errorf("Could not read local data file %s: %v", filename, err)
+	}
+	buf = bytes.NewBuffer(b)
+	f, _ := os.Stat(filename)
+	f.ModTime()
+
+	err = json.NewDecoder(buf).Decode(&jobs)
+	if err != nil {
+		return nil, time.Time{}, err
+	}
+
+	return jobs, f.ModTime(), nil
+
+}
+
+func downloadJobSummaries(dashboard string, storagePath string) error {
+	url := fmt.Sprintf("https://testgrid.k8s.io/%s/summary", dashboard)
+
+	resp, err := http.Get(url)
+	if err != nil {
+		return err
+	}
+	if resp.StatusCode != 200 {
+		return fmt.Errorf("Non-200 response code fetching job summary: %v", resp)
+	}
+	filename := storagePath + "/" + "\"" + strings.ReplaceAll(url, "/", "-") + "\""
+	f, err := os.Create(filename)
+	if err != nil {
+		return err
+	}
+
+	buf := bytes.NewBuffer([]byte{})
+	io.Copy(buf, resp.Body)
+
+	_, err = f.Write(buf.Bytes())
+	return err
+}
+
+// https://testgrid.k8s.io/redhat-openshift-ocp-release-4.4-informing/table?&show-stale-tests=&tab=release-openshift-origin-installer-e2e-azure-compact-4.4
+
+// https://testgrid.k8s.io/redhat-openshift-ocp-release-4.4-informing#release-openshift-origin-installer-e2e-azure-compact-4.4&show-stale-tests=&sort-by-failures=
+
+func downloadJobDetails(dashboard, jobName, storagePath string) error {
+	url := fmt.Sprintf("https://testgrid.k8s.io/%s/table?&show-stale-tests=&tab=%s&grid=old", dashboard, jobName)
+
+	resp, err := http.Get(url)
+	if err != nil {
+		return err
+	}
+	if resp.StatusCode != 200 {
+		return fmt.Errorf("Non-200 response code fetching job details: %v", resp)
+	}
+
+	filename := storagePath + "/" + "\"" + strings.ReplaceAll(url, "/", "-") + "\""
+	f, err := os.Create(filename)
+	if err != nil {
+		return err
+	}
+
+	buf := bytes.NewBuffer([]byte{})
+	io.Copy(buf, resp.Body)
+
+	_, err = f.Write(buf.Bytes())
+	return err
+
+}
